add AbstractComponent
