+++ src/components/board.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createBoardTemplate = () => {
   return (
@@ -6,20 +6,8 @@ const createBoardTemplate = () => {
   );
 };
 
-export default class Board {
-  constructor() {
-    this._element = null;
-  }
+export default class Board extends AbstractComponent {
   getTemplate() {
     return createBoardTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
