+++ src/components/no-tasks.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createNoTasksTemplate = () => {
   return (
@@ -8,20 +8,8 @@ const createNoTasksTemplate = () => {
   );
 };
 
-export default class NoTasks {
-  constructor() {
-    this._element = null;
-  }
+export default class NoTasks extends AbstractComponent {
   getTemplate() {
     return createNoTasksTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
