+++ src/components/sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createSortTemplate = () => {
   return (`<div class="board__filter-list">
@@ -9,20 +9,8 @@ const createSortTemplate = () => {
 `);
 };
 
-export default class Sort {
-  constructor() {
-    this._element = null;
-  }
+export default class Sort extends AbstractComponent {
   getTemplate() {
     return createSortTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
