+++ src/components/task-edit.js
@@ -1,5 +1,6 @@
 import {MONTH_NAMES, DAYS_OF_THE_WEEK, STATUS_COLORS} from '../const.js';
-import {formatTime, createElement} from '../util.js';
+import {formatTime} from '../utils/utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createColorsMarkup = (colors, currentColor) => {
   return colors.map((color, index) => {
@@ -121,21 +122,12 @@ const createTaskEditTemplate = (task) => {
   );
 };
 
-export default class TaskEdit {
+export default class TaskEdit extends AbstractComponent {
   constructor(task) {
+    super();
     this._task = task;
-    this._element = null;
   }
   getTemplate() {
     return createTaskEditTemplate(this._task);
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
