+++ src/components/task.js
@@ -1,5 +1,6 @@
 import {MONTH_NAMES} from '../const.js';
-import {formatTime, createElement} from '../util.js';
+import {formatTime} from '../utils/utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createTaskTemplate = (task) => {
   const {description, dueDate, color, repeatingDays, isArchive, isFavorite} = task;
@@ -61,21 +62,12 @@ const createTaskTemplate = (task) => {
   );
 };
 
-export default class Task {
+export default class Task extends AbstractComponent {
   constructor(task) {
+    super();
     this._task = task;
-    this._element = null;
   }
   getTemplate() {
     return createTaskTemplate(this._task);
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
