+++ src/components/tasks.js
@@ -1,4 +1,4 @@
-import {createElement} from "../util.js";
+import AbstractComponent from './abstract-component.js';
 
 const createTasksTemplate = () => {
   return (
@@ -6,20 +6,8 @@ const createTasksTemplate = () => {
   );
 };
 
-export default class Tasks {
-  constructor() {
-    this._element = null;
-  }
+export default class Tasks extends AbstractComponent {
   getTemplate() {
     return createTasksTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
