+++ src/utils/utils.js
@@ -31,35 +31,9 @@ const getRandomDate = () => {
   return targetDate;
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
 export {
   getRandomInteger,
   getRandomEl,
   formatTime,
   getRandomDate,
-  createElement,
-  RenderPosition,
-  render
 };
